fastapi/pyproject.toml
[tool.poetry]
name = "metaqs-api"
version = "0.1.0"
description = ""
authors = []

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.70.0"
gunicorn = "^20.1.0"
uvicorn = "^0.15.0"
elasticsearch = {extras = ["async"], version = "^7.14.1"}
elasticsearch-dsl = "^7.4.0"
SQLAlchemy = "^1.4.23"
SQLAlchemy-Utils = "^0.37.8"
databases = {extras = ["postgresql"], version = "^0.5.1"}
asyncpg = "^0.24.0"
psycopg2-binary = "^2.9.1"
glom = "^20.11.0"
more-itertools = "^8.8.0"
aioredis = "^2.0.0"
python-slugify = "^5.0.2"
starlette-context = "^0.3.3"
radon = "^5.1.0"
httpx = {version = "^1.0.0*", allow-prereleases = true}
jsonrpcclient = "^4.0.1"
pyLanguagetool = "^0.9.2"
polling2 = "^0.5.0"
fastapi-utils = "^0.2.1"

[tool.poetry.dev-dependencies]
pre-commit = "^2.18.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.vulture]
exclude = []
ignore_decorators = ["@app.route", "@require_*"]
ignore_names = []
make_whitelist = true
min_confidence = 80
paths = ["app"]
sort_by_size = true
verbose = false
